report: Skip data item if version doesn't exist
It's possible for some data from ncm-api to come back
as malformed. We do not want to include it in the report.

This allows `ncm-cli report` to return data from almost all
modules, instead of a single bad module in ncm-api causing the
entire report to fail.

If the version field is falsey then skip this item and drop it
from the report.