report: handle data better if version doesn't exist
It's possible for some data from ncm-api to come back
as malformed. This attempts to fix the report in such cases.

Instead of skipping a module that comes back with unknown
data from ncm-api we set the version to a valid Semver object
aka `0.0.0-unknown-version`.

Additional fixes by: Kevin Swiber <kswiber@gmail.com>
Fixes: https://github.com/nodesource/ncm-cli/issues/155
PR-URL: https://github.com/nodesource/ncm-cli/pull/151